Aa user screen

+++ app/src/main/java/ru/anlyashenko/atmosphereapp/feature/profile/models/DailyMoodStat.kt
@@ -4,6 +4,6 @@ import androidx.compose.ui.graphics.Color
 
 data class DailyMoodStat(
     val dayName: String,
-    val level: Int,
+    val level: Float,
     val color: Color
 )
